Add support for events.

+++ pde1dmex/MexInterface.cpp
@@ -27,8 +27,8 @@ MexInterface::~MexInterface()
 
 void MexInterface::print(const mxArray *a, const char *name)
 {
-  int m = mxGetM(a);
-  int n = mxGetN(a);
+  size_t m = mxGetM(a);
+  size_t n = mxGetN(a);
   const Eigen::Map<Eigen::MatrixXd> A(mxGetPr(a), m, n);
   mexPrintf("%s(%d,%d)\n", name, m, n);
   for (int i = 0; i < m; i++) {
@@ -54,12 +54,12 @@ void MexInterface::callMatlab(const mxArray *inArgs[], int nargin,
     mxArray *a = matOutArgs[i];
     if (!a)
       mexErrMsgIdAndTxt("bvp1d:mexCallMATLAB:arg", "Error in mexCallMATLAB arg.");
-    int retLen = mxGetNumberOfElements(a);
-    int exLen = outArgs[i]->size();
+    size_t retLen = mxGetNumberOfElements(a);
+    size_t exLen = outArgs[i]->size();
     if (retLen != exLen) {
       char msg[1024];
       std::string funcName = getFuncNameFromHandle(inArgs[0]);
-      int m = mxGetM(a), n = mxGetN(a);
+      size_t m = mxGetM(a), n = mxGetN(a);
       sprintf(msg, "In the call to user-defined function:\n\"%s\"\n"
         "returned entry %d had size (%d x %d) but a vector of size (%d x 1)"
         " was expected.", funcName.c_str(), i + 1, m, n, exLen);
@@ -88,8 +88,8 @@ std::string MexInterface::getFuncNameFromHandle(const mxArray *fh)
 
 Eigen::MatrixXd MexInterface::fromMxArray(const mxArray *a)
 {
-  const int m = mxGetM(a);
-  const int n = mxGetN(a);
+  const size_t m = mxGetM(a);
+  const size_t n = mxGetN(a);
   Eigen::MatrixXd mat(m, n);
   std::copy_n(mxGetPr(a), m*n, mat.data());
   return mat;
@@ -97,8 +97,8 @@ Eigen::MatrixXd MexInterface::fromMxArray(const mxArray *a)
 
 Eigen::VectorXd MexInterface::fromMxArrayVec(const mxArray *a)
 {
-  const int m = mxGetM(a);
-  const int n = mxGetN(a);
+  const size_t m = mxGetM(a);
+  const size_t n = mxGetN(a);
   Eigen::VectorXd mat(m*n);
   std::copy_n(mxGetPr(a), m*n, mat.data());
   return mat;
